Fix for React SyntheticEvent & "input" event type
React event handlers do not get called with an instance of js/Event,
they get called with a SyntheticEvent, so `(instance? js/Event value)`
is not a valid check.

Additionally, some genius decided that if you register an `:on-change`
handler, it's perfectly reasonable that the event type you get is
"input", not "change", so we need to make sure we extract that too.