debug-view: add :keys; observe db val, not atom
The debug-view's db previously contained the debuggee's db atom. Now,
instead, it contains the debuggee's db value, and you need to set up a
watch on the debuggee's db atom to `swap!` the debug-view's db when the
debuggee's db changes.  This is to allow the debug view to be mounted
properly in React so that it only re-renders when its db value has
changed (as for any other view).