Reworked view rendering to play nicely with React
The view tree is now transformed at render time into what amounts to a
tree of view functions and the arguments they should be called with. By
introducing a `ViewRenderer` protocol, we allow the actual function
invocation to be controlled externally, so that we can produce React
components which don't re-render unless the arguments to functions in
the relevant branch of the view tree have changed.

Working, but definitely further work is required/desirable.