WIP 0.1.0: introduce :outputs and :transition-fn
Move away from side-effects in transition functions: produce :outputs in
the return value instead, with an output-fn supplied to
consume-dispatch-queue to handle the outputs and make the actual AJAX
calls (for example). This allows you to re-use the same machine with
different output-fns to support, for example, one which makes AJAX calls
to post state to a server, and another which writes to a local database
to store state in a React Native app. This also happens to improve
testability of the application: you can assert against :outputs, now,
rather than having to put something in the DB to affect the behaviour of
your transition functions.

Simplify the `defmachine` form by removing the transition functions from
the transition specification into a separate :transition-fn supplied at
the machine level (intended to allow for a multimethod implementation).