Rename transition -> input; remove dispatch-queue
The notion of a dispatch queue is no longer necessary: now that
`:outputs` are handled separately from transitions, such that the
dispatcher isn't needed during state machine processing at all, there's
no reason to create a separate queue atom and watch it, etc.

With that change, anything related to dispatch has now moved to a
dispatch namespace, since it's no longer core to the operation of the
state machines.