Tests: input responses to :outputs asynchronously
When a test has an :output which is auto-handled, better to delay the
invocation until after the `(input! ...)` call is finished.  In the real
world, :outputs are expected to do their thing asynchronously, but it's
easier to mock them out synchronously in testing.  This way, we can
implement them the easy way in tests, but still get the real-world
asynchronous behaviour.